backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-decouple==3.8
requests==2.31.0
django-ratelimit==4.1.0
openpyxl==3.1.2
